[general] Provide base classes to improve reusability Implements and closes #24
