+++ django_netjsongraph/api/urls.py
@@ -1,9 +1,4 @@
-from django.conf.urls import url
-
 from . import views
+from ..utils import get_api_urls
 
-urlpatterns = [
-    url(r'^topology/$', views.network_collection, name='network_collection'),
-    url(r'^topology/(?P<pk>[^/]+)/$', views.network_graph, name='network_graph'),
-    url(r'^receive/(?P<pk>[^/\?]+)/$', views.receive_topology, name='receive_topology'),
-]
+urlpatterns = get_api_urls(views)
